+++ google/cloud/bigtable/data/_read_rows_state_machine.py
@@ -14,35 +14,14 @@
 #
 from __future__ import annotations
 
-from typing import (
-    List,
-    Any,
-    AsyncIterable,
-    AsyncIterator,
-    AsyncGenerator,
-    Iterator,
-    Callable,
-    Awaitable,
-    Type,
-)
-
-import asyncio
-from functools import partial
-from grpc.aio import RpcContext
+from typing import Type
 
 from google.cloud.bigtable_v2.types.bigtable import ReadRowsResponse
-from google.cloud.bigtable_v2.services.bigtable.async_client import BigtableAsyncClient
 from google.cloud.bigtable.row import Row, Cell, _LastScannedRow
 from google.cloud.bigtable.exceptions import InvalidChunk
-from google.cloud.bigtable.exceptions import _RowSetComplete
-from google.api_core import retry_async as retries
-from google.api_core import exceptions as core_exceptions
-from google.cloud.bigtable._helpers import _make_metadata
-from google.cloud.bigtable._helpers import _attempt_timeout_generator
 
 """
-This module provides a set of classes for merging ReadRowsResponse chunks
-into Row objects.
+This module provides classes for the read_rows state machine:
 
 - ReadRowsOperation is the highest level class, providing an interface for asynchronous
   merging end-to-end
@@ -54,253 +33,6 @@ into Row objects.
   to do on the next chunk.
 - RowBuilder is used by the StateMachine to build a Row object.
 """
-
-
-class _ReadRowsOperation(AsyncIterable[Row]):
-    """
-    ReadRowsOperation handles the logic of merging chunks from a ReadRowsResponse stream
-    into a stream of Row objects.
-
-    ReadRowsOperation.merge_row_response_stream takes in a stream of ReadRowsResponse
-    and turns them into a stream of Row objects using an internal
-    StateMachine.
-
-    ReadRowsOperation(request, client) handles row merging logic end-to-end, including
-    performing retries on stream errors.
-    """
-
-    def __init__(
-        self,
-        request: dict[str, Any],
-        client: BigtableAsyncClient,
-        *,
-        operation_timeout: float = 600.0,
-        per_request_timeout: float | None = None,
-    ):
-        """
-        Args:
-          - request: the request dict to send to the Bigtable API
-          - client: the Bigtable client to use to make the request
-          - operation_timeout: the timeout to use for the entire operation, in seconds
-          - per_request_timeout: the timeout to use when waiting for each individual grpc request, in seconds
-                If not specified, defaults to operation_timeout
-        """
-        self._last_emitted_row_key: bytes | None = None
-        self._emit_count = 0
-        self._request = request
-        self.operation_timeout = operation_timeout
-        # use generator to lower per-attempt timeout as we approach operation_timeout deadline
-        attempt_timeout_gen = _attempt_timeout_generator(
-            per_request_timeout, operation_timeout
-        )
-        row_limit = request.get("rows_limit", 0)
-        # lock in paramters for retryable wrapper
-        self._partial_retryable = partial(
-            self._read_rows_retryable_attempt,
-            client.read_rows,
-            attempt_timeout_gen,
-            row_limit,
-        )
-        predicate = retries.if_exception_type(
-            core_exceptions.DeadlineExceeded,
-            core_exceptions.ServiceUnavailable,
-            core_exceptions.Aborted,
-        )
-
-        def on_error_fn(exc):
-            if predicate(exc):
-                self.transient_errors.append(exc)
-
-        retry = retries.AsyncRetry(
-            predicate=predicate,
-            timeout=self.operation_timeout,
-            initial=0.01,
-            multiplier=2,
-            maximum=60,
-            on_error=on_error_fn,
-            is_stream=True,
-        )
-        self._stream: AsyncGenerator[Row, None] | None = retry(
-            self._partial_retryable
-        )()
-        # contains the list of errors that were retried
-        self.transient_errors: List[Exception] = []
-
-    def __aiter__(self) -> AsyncIterator[Row]:
-        """Implements the AsyncIterable interface"""
-        return self
-
-    async def __anext__(self) -> Row:
-        """Implements the AsyncIterator interface"""
-        if self._stream is not None:
-            return await self._stream.__anext__()
-        else:
-            raise asyncio.InvalidStateError("stream is closed")
-
-    async def aclose(self):
-        """Close the stream and release resources"""
-        if self._stream is not None:
-            await self._stream.aclose()
-        self._stream = None
-        self._emitted_seen_row_key = None
-
-    async def _read_rows_retryable_attempt(
-        self,
-        gapic_fn: Callable[..., Awaitable[AsyncIterable[ReadRowsResponse]]],
-        timeout_generator: Iterator[float],
-        total_row_limit: int,
-    ) -> AsyncGenerator[Row, None]:
-        """
-        Retryable wrapper for merge_rows. This function is called each time
-        a retry is attempted.
-
-        Some fresh state is created on each retry:
-          - grpc network stream
-          - state machine to hold merge chunks received from stream
-        Some state is shared between retries:
-          - _last_emitted_row_key is used to ensure that
-            duplicate rows are not emitted
-          - request is stored and (potentially) modified on each retry
-        """
-        if self._last_emitted_row_key is not None:
-            # if this is a retry, try to trim down the request to avoid ones we've already processed
-            try:
-                self._request["rows"] = _ReadRowsOperation._revise_request_rowset(
-                    row_set=self._request.get("rows", None),
-                    last_seen_row_key=self._last_emitted_row_key,
-                )
-            except _RowSetComplete:
-                # if there are no rows left to process, we're done
-                # This is not expected to happen often, but could occur if
-                # a retry is triggered quickly after the last row is emitted
-                return
-            # revise next request's row limit based on number emitted
-            if total_row_limit:
-                new_limit = total_row_limit - self._emit_count
-                if new_limit == 0:
-                    # we have hit the row limit, so we're done
-                    return
-                elif new_limit < 0:
-                    raise RuntimeError("unexpected state: emit count exceeds row limit")
-                else:
-                    self._request["rows_limit"] = new_limit
-        metadata = _make_metadata(
-            self._request.get("table_name", None),
-            self._request.get("app_profile_id", None),
-        )
-        new_gapic_stream: RpcContext = await gapic_fn(
-            self._request,
-            timeout=next(timeout_generator),
-            metadata=metadata,
-        )
-        try:
-            state_machine = _StateMachine()
-            stream = _ReadRowsOperation.merge_row_response_stream(
-                new_gapic_stream, state_machine
-            )
-            # run until we get a timeout or the stream is exhausted
-            async for new_item in stream:
-                if (
-                    self._last_emitted_row_key is not None
-                    and new_item.row_key <= self._last_emitted_row_key
-                ):
-                    raise InvalidChunk("Last emitted row key out of order")
-                # don't yeild _LastScannedRow markers; they
-                # should only update last_seen_row_key
-                if not isinstance(new_item, _LastScannedRow):
-                    yield new_item
-                    self._emit_count += 1
-                self._last_emitted_row_key = new_item.row_key
-                if total_row_limit and self._emit_count >= total_row_limit:
-                    return
-        except (Exception, GeneratorExit) as exc:
-            # ensure grpc stream is closed
-            new_gapic_stream.cancel()
-            raise exc
-
-    @staticmethod
-    def _revise_request_rowset(
-        row_set: dict[str, Any] | None,
-        last_seen_row_key: bytes,
-    ) -> dict[str, Any]:
-        """
-        Revise the rows in the request to avoid ones we've already processed.
-
-        Args:
-          - row_set: the row set from the request
-          - last_seen_row_key: the last row key encountered
-        Raises:
-          - _RowSetComplete: if there are no rows left to process after the revision
-        """
-        # if user is doing a whole table scan, start a new one with the last seen key
-        if row_set is None or (
-            len(row_set.get("row_ranges", [])) == 0
-            and len(row_set.get("row_keys", [])) == 0
-        ):
-            last_seen = last_seen_row_key
-            return {
-                "row_keys": [],
-                "row_ranges": [{"start_key_open": last_seen}],
-            }
-        # remove seen keys from user-specific key list
-        row_keys: list[bytes] = row_set.get("row_keys", [])
-        adjusted_keys = [k for k in row_keys if k > last_seen_row_key]
-        # adjust ranges to ignore keys before last seen
-        row_ranges: list[dict[str, Any]] = row_set.get("row_ranges", [])
-        adjusted_ranges = []
-        for row_range in row_ranges:
-            end_key = row_range.get("end_key_closed", None) or row_range.get(
-                "end_key_open", None
-            )
-            if end_key is None or end_key > last_seen_row_key:
-                # end range is after last seen key
-                new_range = row_range.copy()
-                start_key = row_range.get("start_key_closed", None) or row_range.get(
-                    "start_key_open", None
-                )
-                if start_key is None or start_key <= last_seen_row_key:
-                    # replace start key with last seen
-                    new_range["start_key_open"] = last_seen_row_key
-                    new_range.pop("start_key_closed", None)
-                adjusted_ranges.append(new_range)
-        if len(adjusted_keys) == 0 and len(adjusted_ranges) == 0:
-            # if the query is empty after revision, raise an exception
-            # this will avoid an unwanted full table scan
-            raise _RowSetComplete()
-        return {"row_keys": adjusted_keys, "row_ranges": adjusted_ranges}
-
-    @staticmethod
-    async def merge_row_response_stream(
-        response_generator: AsyncIterable[ReadRowsResponse],
-        state_machine: _StateMachine,
-    ) -> AsyncGenerator[Row, None]:
-        """
-        Consume chunks from a ReadRowsResponse stream into a set of Rows
-
-        Args:
-          - response_generator: AsyncIterable of ReadRowsResponse objects. Typically
-                this is a stream of chunks from the Bigtable API
-        Returns:
-            - AsyncGenerator of Rows
-        Raises:
-            - InvalidChunk: if the chunk stream is invalid
-        """
-        async for row_response in response_generator:
-            # unwrap protoplus object for increased performance
-            response_pb = row_response._pb
-            last_scanned = response_pb.last_scanned_row_key
-            # if the server sends a scan heartbeat, notify the state machine.
-            if last_scanned:
-                yield state_machine.handle_last_scanned_row(last_scanned)
-            # process new chunks through the state machine.
-            for chunk in response_pb.chunks:
-                complete_row = state_machine.handle_chunk(chunk)
-                if complete_row is not None:
-                    yield complete_row
-            # TODO: handle request stats
-        if not state_machine.is_terminal_state():
-            # read rows is complete, but there's still data in the merger
-            raise InvalidChunk("read_rows completed with partial state remaining")
 
 
 class _StateMachine:
@@ -579,7 +311,7 @@ class _RowBuilder:
         self.current_key: bytes | None = None
         self.working_cell: Cell | None = None
         self.working_value: bytearray | None = None
-        self.completed_cells: List[Cell] = []
+        self.completed_cells: list[Cell] = []
 
     def start_row(self, key: bytes) -> None:
         """Called to start a new row. This will be called once per row"""
@@ -590,7 +322,7 @@ class _RowBuilder:
         family: str,
         qualifier: bytes,
         timestamp_micros: int,
-        labels: List[str],
+        labels: list[str],
     ) -> None:
         """called to start a new cell in a row."""
         if self.current_key is None:
